requirements.txt
Django==3.2
pycodestyle==2.7.0
django-bootstrap-modal-forms==2.1.0
django-widget-tweaks==1.4.8
Celery==5.1.2
psycopg2==2.9.1
django-sendgrid-v5==0.9.1